Consolidate all of the base display wraps
